the endpoints work but 50-50
- changed the types of the IDs to int

+++ src/main/java/com/example/twittertrial/DTO/CommentDto.java
@@ -1,16 +1,16 @@
 package com.example.twittertrial.DTO;
 
 public class CommentDto {
-    private Long commentID;
+    private int commentID;
     private String commentBody;
-    private Long postID;
-    private Long userID; // User ID instead of User object
+    private int postID;
+    private int userID; // User ID instead of User object
 
-    public Long getCommentID() {
+    public int getCommentID() {
         return commentID;
     }
 
-    public void setCommentID(Long commentID) {
+    public void setCommentID(int commentID) {
         this.commentID = commentID;
     }
 
@@ -22,23 +22,23 @@ public class CommentDto {
         this.commentBody = commentBody;
     }
 
-    public Long getPostID() {
+    public int getPostID() {
         return postID;
     }
 
-    public void setPostId(Long postID) {
+    public void setPostId(int postID) {
         this.postID = postID;
     }
 
-    public Long getUserID() {
+    public int getUserID() {
         return userID;
     }
 
-    public void setUserID(Long userID) {
+    public void setUserID(int userID) {
         this.userID = userID;
     }
 
-    public CommentDto(Long commentID, String commentBody, Long postID, Long userID) {
+    public CommentDto(int commentID, String commentBody, int postID, int userID) {
         this.commentID = commentID;
         this.commentBody = commentBody;
         this.postID = postID;
