+++ src/main/java/com/example/twittertrial/Entity/User.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String email;
     private String name;
@@ -25,11 +25,11 @@ public class User {
 
     }
 
-    public Long getID() {
+    public int getID() {
         return id;
     }
 
-    public void setID(Long id) {
+    public void setID(int id) {
         this.id = id;
     }
 
@@ -57,24 +57,4 @@ public class User {
         this.password = password;
     }
 
-//    // toString method
-//    @Override
-//    public String toString() {
-//        return "User{" +
-//                "id=" + id +
-//                ", email='" + email + '\'' +
-//                ", name='" + name + '\'' +
-//                ", password='" + password + '\'' +
-//                '}';
-//    }
-//
-//    // equals and hashcode methods
-//    @Override
-//
-//    public boolean equals(Object o) {
-//        if (this == o) return true;
-//        if (o == null || getClass() != o.getClass()) return false;
-//        User user = (User) o;
-//        return id.equals(user.id) && email.equals(user.email) && name.equals(user.name) && password.equals(user.password);
-//    }
 }
