+++ src/main/java/com/example/twittertrial/Repository/UserRepository.java
@@ -1,10 +1,11 @@
 package com.example.twittertrial.Repository;
 
 import com.example.twittertrial.Entity.User;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
